Increase intervals for session and timer handlers
Increased the intervals for session and timer handlers from 1 second to 10 seconds in dsu-fabric and demiurge-wallet. The sessionHandler function was also changed from a global to a local constant. This optimizes the performance and reduces the number of calls to the server, while also improving scope control in the code.